backend/requirements.txt
Django==4.0
drf-spectacular==0.26.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.6
python-dotenv==0.21.0
gunicorn==20.0.4
six==1.16.0
pytz
web3